# openapi.yaml
openapi: 3.0.0
info:
  title: BankWiz API
  version: 0.1.0
  description: An API for accessing banking functionalities provided by BankWiz
servers:
  - url: 'http://localhost:8080/api'
...
paths:
...
  /user:
    get:
      tags:
        - User
      summary: Get the current user information
      operationId: getCurrentUserInfo
      responses:
        '200':
          description: Successfully retrieved user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Invalid request. Please check the provided data.
    put:
      tags:
        - User
      summary: Update user information
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User information updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Invalid request. Please check the provided data.
  /user/users:
    get:
      tags:
        - User
      summary: Get a list of users
      operationId: getUsers
      responses:
        '200':
          description: Successfully retrieved the list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
        '400':
          description: Invalid request. Please check the provided data.
  '/user/{userId}':
    get:
      tags:
        - User
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      summary: Get user information
      operationId: getUser
      responses:
        '200':
          description: Successfully retrieved user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Invalid request. Please check the provided data.
...
components:
  schemas:
...
    UserUpdateRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
...
    UserDTO:
      type: object
      properties:
        userId:
          type: integer
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: johndoe@example.com
      required:
        - userId
        - firstName
        - lastName
        - email
...